# Conflicts:
#	CMakeLists.txt
#	Common/CMakeLists.txt
#	Common/core/def_version.h
#	Common/game/main_game_file.cpp
#	Common/game/room_file.cpp
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Editor/AGS.Types/SerializeUtils.cs
#	Engine/ac/character.cpp
#	Engine/ac/global_viewframe.cpp
#	Engine/ac/viewframe.cpp
#	Engine/media/audio/audio.cpp
#	Engine/platform/windows/gfx/ali3dd3d.cpp
#	Solutions/Common.Lib/Common.Lib.vcxproj
#	Solutions/Common.Lib/Common.Lib.vcxproj.filters
#	version.json

+++ Engine/game/game_init.cpp
@@ -394,7 +394,7 @@ HGameInitError InitGameState(const LoadedGameEntities &ents, GameDataVersion dat
     if (game.numfonts == 0)
         return new GameInitError(kGameInitErr_NoFonts);
     if (game.audioClipTypes.size() > MAX_AUDIO_TYPES)
-        return new GameInitError(kGameInitErr_TooManyAudioTypes, String::FromFormat("Required: %u, max: %d", game.audioClipTypes.size(), MAX_AUDIO_TYPES));
+        return new GameInitError(kGameInitErr_TooManyAudioTypes, String::FromFormat("Required: %zu, max: %zu", game.audioClipTypes.size(), (size_t)MAX_AUDIO_TYPES));
 
     //
     // CLNUP: this stage is removed
